Add an event loop for the SMS application layer
This commit hasn't been fully tested, but it compiled in Rust and isn't
that the same thing? As part of refocusing on trying to get the GSM
module to a bare minimum level of working, I added a poor event loop
into the SMS manager to fetch new messages and respond to requests
to read those messages (and send SMS messages). Next step here is to
hook this mechanism into the HTTP server for GETing new messages and
POSTing messages to send.